+++ veilid-core/src/routing_table/tests/mod.rs
@@ -8,7 +8,7 @@ pub(crate) mod mock_registry {
     pub(crate) async fn init<S: AsRef<str>>(namespace: S) -> VeilidComponentRegistry {
         let (update_callback, config_callback) = setup_veilid_core_with_namespace(namespace);
         let veilid_config =
-            VeilidConfig::new_from_callback(config_callback, update_callback).unwrap();
+            VeilidStartupOptions::new_from_callback(config_callback, update_callback).unwrap();
         let registry = VeilidComponentRegistry::new(veilid_config);
         registry.enable_mock();
         registry.register(ProtectedStore::new);